+++ src/main/java/org/folio/marc/migrations/controllers/delegates/MarcMigrationsService.java
@@ -1,6 +1,7 @@
 package org.folio.marc.migrations.controllers.delegates;
 
 import java.util.UUID;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.marc.migrations.controllers.mappers.MarcMigrationMapper;
 import org.folio.marc.migrations.domain.dto.EntityType;
@@ -8,29 +9,28 @@ import org.folio.marc.migrations.domain.dto.MigrationOperation;
 import org.folio.marc.migrations.domain.dto.NewMigrationOperation;
 import org.folio.marc.migrations.domain.dto.OperationType;
 import org.folio.marc.migrations.exceptions.ApiValidationException;
+import org.folio.marc.migrations.services.MigrationOrchestrator;
 import org.folio.marc.migrations.services.operations.OperationsService;
 import org.folio.spring.exception.NotFoundException;
 import org.springframework.stereotype.Service;
 
 @Log4j2
 @Service
+@RequiredArgsConstructor
 public class MarcMigrationsService {
 
   static final String NOT_FOUND_MSG = "MARC migration operation was not found [id: %s]";
 
   private final MarcMigrationMapper mapper;
   private final OperationsService operationsService;
-
-  public MarcMigrationsService(MarcMigrationMapper mapper, OperationsService operationsService) {
-    this.mapper = mapper;
-    this.operationsService = operationsService;
-  }
+  private final MigrationOrchestrator migrationOrchestrator;
 
   public MigrationOperation createNewMigration(NewMigrationOperation newMigrationOperation) {
     log.debug("createNewMigration::Trying to create new migration operation: {}", newMigrationOperation);
     validate(newMigrationOperation);
     var operation = mapper.toEntity(newMigrationOperation);
     var newOperation = operationsService.createOperation(operation);
+    migrationOrchestrator.submitMappingTask(newOperation);
     return mapper.toDto(newOperation);
   }